+++ src/components/mypage/Post.tsx
@@ -30,7 +30,7 @@ const Post = (props: Props) => {
         {/*best 태그, 해설게시판 태그*/}
         {topElement ? topElement : null}
         <div className={'flex gap-x-3 items-center justify-between'}>
-          <div className={'flex flex-col gap-y-1'}>
+          <div className={'flex flex-col gap-y-1 w-[60%]'}>
             <div className={'text-h4 font-semibold truncate'}>{title}</div>
             <div className={'text-h4 font-normal line-clamp-2'}>{content}</div>
             {/*좋아요 수, 댓글 수*/}
