+++ src/lib/hooks/useGetTotalSearchResults.tsx
@@ -10,16 +10,7 @@ const getKey = (
   postType: BoardType,
   certificateId: number,
   sortField: string,
-  keyword?: string,
 ) => {
-  //검색 키워드가 있고, boardType 이 REVIEW 가 아니면, 첫페이지
-  if (size === 0 && postType !== 'REVIEW') {
-    return `/certificates/${certificateId}/search?postType=${postType}&keyword=${keyword}&page=${size}&size=10&sortFields=${sortField}`;
-  }
-  //검색 키워드가 있고, boardType 이 REVIEW 가 아니면, 첫 페이지가 아닐 때
-  if (previousPageData && !previousPageData.result.hasNext && postType !== 'REVIEW') {
-    return `/certificates/${certificateId}/search?postType=${postType}&keyword=${keyword}&page=${size}&size=10&sortFields=${sortField}`;
-  }
   //검색 키워드가 없고, boardType 이 REVIEW 가 아니면, 첫페이지
   if (size === 0 && postType !== 'REVIEW') {
     return `/certificates/${certificateId}/search?postType=${postType}&page=${size}&size=10&sortFields=${sortField}`;
@@ -32,21 +23,15 @@ const getKey = (
     return null;
   }
 };
-const useGetTotalSearchResults = (
-  postType: BoardType,
-  certificateId: number,
-  sortField: string,
-  keyword?: string,
-) => {
+const useGetTotalSearchResults = (postType: BoardType, certificateId: number, sortField: string) => {
   const { data, isLoading, error, size, setSize, mutate } = useSWRInfinite<AxiosResponse<ResponsePostType>>(
-    (pageIndex, previousPageData) => getKey(pageIndex, previousPageData, postType, certificateId, sortField, keyword, isOnKeyDown),
+    (pageIndex, previousPageData) => getKey(pageIndex, previousPageData, postType, certificateId, sortField),
     swrGetFetcher,
     {
       revalidateAll: true,
     },
   );
 
-  const parseResultList = data ? data.map((item) => item).flat() : [];
 
   return {
     userPostsList: data ? data : [],
